Create launch files for 1394 cameras.
Uses either of two ROS 1394 camera drivers:

  cameradc1394 (from cmu-ros-pkg)
  camera1394 (Patrick Beeson's port of the Player camera driver to ROS).  

At least one camera driver must be installed. For now, to avoid
confusion, the art_image package does not explicitly depend on
either. When we settle on one solution, it will.

Add example YAML file (for Jack O'Quin's Unibrain Fire-i camera).  The
GUID of the camera is used for the file name.